+++ x-pack/plugins/observability_solution/apm/server/routes/entities/services/get_service_entity_summary.ts
@@ -5,10 +5,10 @@
  * 2.0.
  */
 
-import { EntitiesESClient } from '../../lib/helpers/create_es_client/create_entities_es_client/create_entities_es_client';
-import { withApmSpan } from '../../utils/with_apm_span';
-import { getServiceLatestEntity } from './get_service_latest_entity';
-import { ServiceEntities } from './types';
+import type { EntitiesESClient } from '../../../lib/helpers/create_es_client/create_entities_es_client/create_entities_es_client';
+import { withApmSpan } from '../../../utils/with_apm_span';
+import { getServiceLatestEntity } from '../get_service_latest_entity';
+import type { ServiceEntities } from '../types';
 
 interface Params {
   entitiesESClient: EntitiesESClient;
