+++ x-pack/plugins/observability_solution/apm/server/routes/services/route.ts
@@ -78,7 +78,7 @@ import {
   ServiceTransactionTypesResponse,
 } from './get_service_transaction_types';
 import { getThroughput, ServiceThroughputResponse } from './get_throughput';
-import { getServiceEntitySummary } from '../entities/get_service_entity_summary';
+import { getServiceEntitySummary } from '../entities/services/get_service_entity_summary';
 import { ENVIRONMENT_ALL } from '../../../common/environment_filter_values';
 import { createEntitiesESClient } from '../../lib/helpers/create_es_client/create_entities_es_client/create_entities_es_client';
